+++ R/countrycode.R
@@ -101,14 +101,16 @@ countrycode <- function(sourcevar, origin, destination, warn = TRUE, nomatch = N
     if (is.null(custom_dict)) {
         dictionary <- countrycode::codelist
         attr(dictionary, "origin_valid") <- c(
-          "cctld", "country.name", "country.name.de", "cowc", "cown", "dhs",
-          "ecb", "eurostat", "fao", "fips", "gaul", "genc2c", "genc3c",
-          "genc3n", "gwc", "gwn", "imf", "ioc", "iso2c", "iso3c", "iso3n",
-          "p4c", "p4n", "un", "un_m49", "unicode.symbol", "unhcr", "unpd", "vdem", 
-          "wb", "wb_api2c", "wb_api3c", "wvs", "country.name.en.regex",
-          "country.name.de.regex")
+            "cctld", "country.name", "country.name.de", "country.name.fr", "country.name.it",
+            "cowc", "cown", "dhs", "ecb", "eurostat", "fao", "fips", "gaul",
+            "genc2c", "genc3c", "genc3n", "gwc", "gwn", "imf", "ioc", "iso2c",
+            "iso3c", "iso3n", "p4c", "p4n", "un", "un_m49", "unicode.symbol",
+            "unhcr", "unpd", "vdem", "wb", "wb_api2c", "wb_api3c", "wvs",
+            "country.name.en.regex", "country.name.de.regex", "country.name.fr.regex", "country.name.it.regex")
         attr(dictionary, "origin_regex") <- c("country.name.de.regex",
-                                              "country.name.en.regex")
+                                              "country.name.en.regex",
+                                              "country.name.fr.regex",
+                                              "country.name.it.regex")
     } else {
         dictionary <- custom_dict
     }
@@ -118,7 +120,7 @@ countrycode <- function(sourcevar, origin, destination, warn = TRUE, nomatch = N
         if (origin == 'country.name') {
             origin <- 'country.name.en'
         }
-        if (origin %in% c('country.name.en', 'country.name.de')) {
+        if (origin %in% c('country.name.en', 'country.name.de', 'country.name.it', 'country.name.fr')) {
             origin <- paste0(origin, '.regex')
         }
         destination[destination == "country.name"] <- 'country.name.en'